fixing cart

+++ app/src/main/java/com/gogoy/components/item/ItemRelatedAdapter.kt
@@ -28,7 +28,7 @@ class ItemRelatedAdapter(
         val item = list[position]
 
         holder.tvName.text = item.name
-        holder.tvPrice.text = toRupiah(item.price.toDouble())
+        holder.tvPrice.text = toRupiah(item.price)
         holder.tvOwner.text = item.owner
         holder.ivBadge.setImageResource(item.badge)
     }
